Stop using ResolverPlugin
We were using it to find main attrs in bower.json files. Trouble is
some modules have both a package.json and a bower.json file, and in
that case bower.json was winning. Often the file pointed to by bower
was a precompiled file. To work around, I added another resolver for
main in package.json, but this overrides the default webpack logic
which takes into account alternate main attrs in package.json like
"browser".

See https://github.com/isaacs/inherits/blob/master/package.json
for an example of this being used.