Fix dev asset json containing absolute URLs
This was preventing the cudl-viewer's cudl.ui.dev.baseUrl setting from
having any effect, as the absolute URL in the dev asset file would
blat the base URL from cudl-viewer when resolving against each other.

The version bump to assets-webpack-plugin also seems to  remove the
useless source map entries in the asset json file which is a bonus.