Work around (chrome?) sourcemap debugging issue
See: https://github.com/webpack/webpack/issues/1487

Line numbers seem to be off in chrome when setting break points in
files which are source mapped.

In dev mode we use eval-source-map which doesn't exhibit the issue. In
production we use regular source-map, as eval-source-map is not
suitable for production use.