Hide the bookmark button if the user isn't logged in
Now that the bookmark option is POST only for CSRF protection, it gets
hit with a GET after a post-login-exchange redirect, which doesn't
work.

Easiest thing for now is to just hide the button for logged out
users, but it'd be nice to improve the functionality of it in general,
e.g. to change its state to be a remove button when the item is already
bookmarked, as well as indicate that bookmarking is available when
logged in.